% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hpo_test.r
\name{backtest}
\alias{backtest}
\title{Backtest}
\usage{
backtest(m1, rt, orig, h, xre = NULL, fixed = NULL, include.mean = TRUE)
}
\arguments{
\item{m1}{a time-series model object.}

\item{rt}{the time series.}

\item{orig}{forecast origin.}

\item{h}{forecast horizon.}

\item{xre}{the independent variables.}

\item{fixed}{parameter constriant.}

\item{inc.mean}{a logicial value for constant term of the model. Default is TRUE.}
}
\value{
The function returns a list with following components:
\item{orig}{the starting forecast origin.}
\item{err}{observed value minus fitted value.}
\item{rmse}{RMSE of out-of-sample forecasts.}
\item{mabso}{mean absolute error of out-of-sample forecasts.}
\item{bias}{bias of out-of-sample foecasts.}
}
\description{
Backtest
}
