% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_SETAR.r
\name{MlmNeSS}
\alias{MlmNeSS}
\title{This program performs multivariate linear regression analysis.}
\usage{
MlmNeSS(y, z, subset = NULL, SD = FALSE, include.mean = TRUE)
}
\description{
This program performs multivariate linear regression analysis.
}
