% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSM.r
\name{MSM.fit}
\alias{MSM.fit}
\title{Fitting Univariate Autoregressive Markov Switching Models}
\usage{
MSM.fit(y, p, nregime = 2, include.mean = T, sw = NULL)
}
\arguments{
\item{y}{a time series}

\item{p}{AR order}

\item{nregime}{the number of regimes}

\item{include.mean}{a logical value for including constant terms}

\item{sw}{logical values for whether coefficients are switching. The length of sw has to be equal of the number of coefficients in the model plus include.mean.}
}
\value{
MSM.fit returns an object of class MSM.lm or MSM.glm, depending on the input model.
}
\description{
Fit autoregressive Markov switching models to a univariate time series using the package MSwM.
}
\examples{
y=MSM.sim(100,c(1,1),0.7,-0.5,c(0.5,0.6),c(1,1),c(0,0),500)
MSM.fit(y$series,1,2,TRUE,c(TRUE,TRUE,TRUE))
}
