% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pie.R
\name{my_pie}
\alias{my_pie}
\title{Using a dataframe specific column to plot a pie figure}
\usage{
my_pie(data, ncol)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{ncol}{specific column, factor}
}
\value{
a figure as ggplot2 object
}
\description{
Using a dataframe specific column to plot a pie figure
}
\examples{
my_pie(mtcars,ncol=2)
my_pie(diamonds,ncol=3)
}
