% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hpo_test.r
\name{PRnd}
\alias{PRnd}
\title{ND Test}
\usage{
PRnd(x, m = 10, p = 0, q = 0)
}
\arguments{
\item{x}{time series}
}
\value{
PRnd function outputs the ND test statistic and its p-value.
}
\description{
Compute the ND test statistic of Pena and Rodriguez (2006, JSPI).
}
\examples{
y=rnorm(100)
PRnd(y,10,0,0)
}
\references{
Pena, D, and Rodriguez, J. (2006) The log of the determinant of the autocorrelation matrix for testing goodness of fit in time series. \emph{Journal of Statistical Planning and Inference}, 136, 2706-2718.
}
