% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test3-package.R
\docType{package}
\name{test3-package}
\alias{test3}
\alias{test3-package}
\title{test3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramnathv/test3}
  \item Report bugs at \url{https://github.com/ramnathv/test3/issues}
}

}
\author{
\strong{Maintainer}: Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

}
\keyword{internal}
