% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvAR.r
\name{tvAR}
\alias{tvAR}
\title{Estimate Time-Varying Coefficient AR Models}
\usage{
tvAR(x, lags = c(1), include.mean = TRUE)
}
\arguments{
\item{x}{a time series of data.}

\item{lags}{the lagged variables used, e.g. lags=c(1,3) means lag-1 and lag-3 are used as regressors. 
It is more flexible than specifying an order.}

\item{include.mean}{a logical value indicating whether the constant terms are included.}
}
\value{
\code{trAR} function returns the value from function \code{dlmMLE}.
}
\description{
Estimate time-varying coefficient AR models.
}
\examples{
x1=rnorm(100)
x2=arima.sim(n = 63, list(ar = c(0.8897, -0.4858)),sd = sqrt(0.1796))
x=c(x1,x2)
est=tvAR(x,2)
}
