% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvAR.r
\name{tvARFiSm}
\alias{tvARFiSm}
\title{Filtering and Smoothing for Time-Varying AR Models}
\usage{
tvARFiSm(x, lags = c(1), include.mean = TRUE, par)
}
\arguments{
\item{x}{a time series of data.}

\item{lags}{the lag of AR order.}

\item{include.mean}{a logical value indicating whether the constant terms are included.}

\item{par}{the fitted time-varying AR models. It can be an object returned by function. \code{tvAR}.}
}
\value{
\code{trARFiSm} function return values returned by function \code{dlmFilter} and \code{dlmSmooth}.
}
\description{
This function performs forward filtering and backward smoothing for a fitted time-varying AR model with parameters in 'par'.
}
\examples{
x1=rnorm(100)
x2=arima.sim(n = 63, list(ar = c(0.8897, -0.4858)),sd = sqrt(0.1796))
x=c(x1,x2)
est=tvAR(x,2)
trARFiSm(x,2,TRUE,est$par)
}
