% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_dummy.r
\name{hfDummy}
\alias{hfDummy}
\title{Create Dummy Variables for High-Frequency iIntraday Seasonality}
\usage{
hfDummy(int = 1, Fopen = 10, Tend = 10, days = 1, pooled = 1,
  skipmin = 0)
}
\arguments{
\item{int}{length of time interval in minutes.}

\item{Fopen}{number of dummies from the market open.}

\item{Tend}{number of dummies to the market close.}

\item{days}{number of trading days in the data.}
}
\description{
Create Dummy Variables for High-Frequency iIntraday Seasonality
}
